renamed services to views
