+++ core/src/main/java/gr17/noodleio/game/views/LobbyService.kt
@@ -1,12 +1,9 @@
-package gr17.noodleio.game.services
+package gr17.noodleio.game.views
 
 import gr17.noodleio.game.config.EnvironmentConfig
 import gr17.noodleio.game.models.Lobby
 import gr17.noodleio.game.models.LobbyPlayer
-import io.github.jan.supabase.postgrest.from
-import io.github.jan.supabase.postgrest.rpc
 import kotlinx.coroutines.runBlocking
-import kotlinx.datetime.Instant
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.buildJsonObject
