+++ core/src/main/java/gr17/noodleio/game/views/RealtimeCursorService.kt
@@ -1,4 +1,4 @@
-package gr17.noodleio.game.services
+package gr17.noodleio.game.views
 
 import gr17.noodleio.game.config.EnvironmentConfig
 import gr17.noodleio.game.models.CursorPosition
