+++ core/src/main/java/gr17/noodleio/game/views/RealtimeGameSessionService.kt
@@ -1,6 +1,4 @@
-package gr17.noodleio.game.services
-
-
+package gr17.noodleio.game.views
 
 /*
 Use database functions for:
@@ -15,12 +13,16 @@ Move to application code:
 
 - Collision detection () => set wasEaten
 - Food spawning (as a scheduled function)
+
+*
 * */
 
-class RealtimePlayerGameStateService {
+
+class RealtimeGameSessionService {
 
     // https://supabase.com/docs/guides/realtime/postgres-changes?queryGroups=language&language=kotlin
 
     // see the proof of concept CursorRealtimeCursorService
+
 
 }
